src/requirements.txt
opensea-api==0.1.7
numpy==1.22.3
pandas==1.3.5
requests==2.27.1
streamlit==1.8.1
json5==0.9.5
psycopg2-binary==2.9.3
matplotlib==3.5.2
plotly==5.8.0
web3==5.29.1